Add data structure for efficient identity matching of fingerprints
