+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: morgancpp
-Version: 0.3.2
+Version: 0.3.3
 Date: 2019-12-08
 License: MIT
 Title: Morgan fingerprints in C++
