+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(MorganFPS)
+export(MorganMap)
 export(tanimoto)
 importFrom(Rcpp,cpp_object_initializer)
 useDynLib(morgancpp)
