+++ R/RcppExports.R
@@ -1,6 +1,13 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+#' @name MorganMap
+#' @title Morgan fingerprint collection for identity checking
+#' @description Efficient structure for checking identity of Morgan fingerprints
+#' @importFrom Rcpp cpp_object_initializer
+#' @export
+NULL
+
 #' @name MorganFPS
 #' @title Morgan fingerprints collection
 #' @description Efficient structure for storing a set of Morgan fingerprints
