+++ R/zzz.R
@@ -2,3 +2,4 @@
 NULL
 
 Rcpp::loadModule("morgan_cpp", TRUE)
+Rcpp::loadModule("morgan_identity_cpp", TRUE)
