+++ man/MorganMap.Rd
@@ -1,0 +1,8 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/RcppExports.R
+\name{MorganMap}
+\alias{MorganMap}
+\title{Morgan fingerprint collection for identity checking}
+\description{
+Efficient structure for checking identity of Morgan fingerprints
+}
